Reduce dataset to a single hash
This cleans up the "dual" source dataset problem, by ensuring one
single ruby hash as the source dataset for all exports. This was
partly done already, but CSV exports were using the LIST so far. This
removes the list entirely, and converts everything to use the HASH
as the source.

With this, patches need to be applied only once, on the HASH and are
much more easier to manage (Since you can do lookups on the HASH
by IFSC)